# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Explore Kerinci API",
        "description": "REST API backend untuk aplikasi pariwisata Explore Kerinci. Gunakan endpoint `/api/auth/login` untuk mendapatkan token, lalu klik tombol **Authorize** dan masukkan token tersebut.",
        "contact": {
            "name": "Explore Kerinci",
            "email": "admin@wisker.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Development Server (php artisan serve)"
        }
    ],
    "paths": {
...
        "/api/destinations": {
            "get": {
                "tags": [
                    "Destinations"
                ],
                "summary": "Daftar destinasi wisata (publik)",
                "operationId": "b77633672e38c28f34c971e03fba189e",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari berdasarkan nama atau deskripsi",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "gunung"
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter berdasarkan slug kategori",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "alam-pegunungan"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nomor halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Destination"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/destinations/{slug}": {
            "get": {
                "tags": [
                    "Destinations"
                ],
                "summary": "Detail destinasi berdasarkan slug (publik)",
                "operationId": "00c53fe70bf27ff4ab9c1fd97c385852",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "gunung-kerinci"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Destination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Destinasi tidak ditemukan"
                    }
                }
            }
        },
        "/api/destinations/{id}/photos": {
            "post": {
                "tags": [
                    "Photos"
                ],
                "summary": "Upload foto untuk destinasi",
                "operationId": "fa9c50237f59ed6f1d9490a41c64989f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "photo"
                                ],
                                "properties": {
                                    "photo": {
                                        "description": "File gambar (max 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "is_primary": {
                                        "description": "Jadikan foto utama",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Foto berhasil diunggah",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/DestinationImage"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Destinasi tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "description": "OpenAPI schema definitions for Explore Kerinci API.\nThis file is scanned by L5-Swagger automatically.",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "budi@example.com"
                    },
                    "avatar_url": {
                        "type": "string",
                        "example": "http://127.0.0.1:8000/storage/avatars/photo.jpg",
                        "nullable": true
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "user",
                            "admin"
                        ],
                        "example": "user"
                    }
                },
                "type": "object"
            },
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Alam Pegunungan"
                    },
                    "slug": {
                        "type": "string",
                        "example": "alam-pegunungan"
                    }
                },
                "type": "object"
            },
            "DestinationImage": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "image_url": {
                        "type": "string",
                        "example": "http://127.0.0.1:8000/storage/destinations/img.jpg"
                    },
                    "is_primary": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "Review": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "rating": {
                        "type": "integer",
                        "maximum": 5,
                        "minimum": 1,
                        "example": 5
                    },
                    "comment": {
                        "type": "string",
                        "example": "Pemandangan sangat indah!"
                    },
                    "approved_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z",
                        "nullable": true
                    },
                    "user": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "Destination": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Gunung Kerinci"
                    },
                    "slug": {
                        "type": "string",
                        "example": "gunung-kerinci"
                    },
                    "description": {
                        "type": "string",
                        "example": "Gunung tertinggi di Sumatera..."
                    },
                    "location": {
                        "type": "string",
                        "example": "Sungai Penuh, Kerinci"
                    },
                    "ticket_price": {
                        "type": "number",
                        "format": "float",
                        "example": 25000
                    },
                    "open_hours": {
                        "type": "string",
                        "example": "06:00 - 18:00"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    },
                    "rating_avg": {
                        "type": "number",
                        "format": "float",
                        "example": 4.8,
                        "nullable": true
                    },
                    "review_count": {
                        "type": "integer",
                        "example": 30
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DestinationImage"
                        }
                    },
                    "reviews": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Review"
                        }
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token Sanctum di sini. Contoh: **Bearer 1|abc...**",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}